code/requirements.txt
pandas
Pillow
reportlab
pywin32
